Support IdP-initiated SLO with CSRF-protected logout
When the Identity Provider sends a single logout request to /saml/logout
and logout is configured as CSRF protected in security.yaml, the request
fails as it is missing the CSRF token.

A new exception listener has been added that is able to detect this
situation, and, upon successful validation of the SAML request, redirect
to the logout URL with a valid CSRF token and the SLO request. The
logout is then processed as usual.